+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+#### Changed
+* **\[processor\]** Remove restriction that methods annotated with `@Callback` must have types that are
+  annotated with `@JsFunction`.
+
 ### [v0.62](https://github.com/react4j/react4j/tree/v0.62) (2018-04-09)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.61...v0.62)